Added changeable views for all planes
