+++ sql/SQL-Setup.sql
@@ -2,9 +2,9 @@
 
 CREATE TABLE Points (
     Points_ID BIGSERIAL PRIMARY KEY,
-    Points_x DOUBLE PRECISION,
-    Points_y DOUBLE PRECISION,
-    Points_z DOUBLE PRECISION,
+    Points_x FLOAT,
+    Points_y FLOAT,
+    Points_z FLOAT,
     Name VARCHAR(100)
 );
 
@@ -12,5 +12,5 @@ CREATE TABLE Lines (
     Lines_ID BIGSERIAL PRIMARY KEY,
     LINES_PointA_Points_ID INTEGER REFERENCES Points (Points_ID),
     Lines_PointB_Points_ID INTEGER REFERENCES Points(Points_ID),
-    Lines_length DOUBLE PRECISION
+    Lines_length FLOAT
 );
