+++ src/View/Orth_YZ_OpenGLWidget.cpp
@@ -1,21 +1,20 @@
 // Orth_XZ_OpenGLWidget.cpp
-#include "Orth_XZ_OpenGLWidget.h"
+#include "Orth_YZ_OpenGLWidget.h"
 #include <QOpenGLFunctions>
 #include <Model.h>
 #include <QDebug>
 #include <cmath>
 
-Orth_XZ_OpenGLWidget::Orth_XZ_OpenGLWidget(Model *model, Controller *controller, QWidget *parent)
+Orth_YZ_OpenGLWidget::Orth_YZ_OpenGLWidget(Model *model, Controller *controller, QWidget *parent)
     : View_OpenGLWidget(model, controller, parent)
 {
 }
 
-Orth_XZ_OpenGLWidget::~Orth_XZ_OpenGLWidget()
+Orth_YZ_OpenGLWidget::~Orth_YZ_OpenGLWidget()
 {
 }
 
-void Orth_XZ_OpenGLWidget::drawGrid() {
-    //TODO: rework, make grid Size changeable and show world coordinates
+void Orth_YZ_OpenGLWidget::drawGrid() {
     float viewWidth = width() / zoomLevel;
     float viewHeight = height() / zoomLevel;
     
@@ -36,60 +35,57 @@ void Orth_XZ_OpenGLWidget::drawGrid() {
     // Vertikale Linien (X-Achse)
     glBegin(GL_LINES);
     for (float x = left; x <= right; x += gridSize) {
-        glVertex3f(x, 0.0f, bottom);  // X-Z Ebene
-        glVertex3f(x, 0.0f, top);     // X-Z Ebene
+        glVertex3f(x, 100.0f, bottom);  // X-Y Ebene
+        glVertex3f(x, 100.0f, top);     // X-Y Ebene
     }
-    // Horizontale Linien (Z-Achse)
-    for (float z = bottom; z <= top; z += gridSize) {
-        glVertex3f(left, 0.0f, z);    // X-Z Ebene
-        glVertex3f(right, 0.0f, z);   // X-Z Ebene
+    // Horizontale Linien (Y-Achse)
+    for (float y = bottom; y <= top; y += gridSize) {
+        glVertex3f(left, y, 100.0f);    // X-Y Ebene
+        glVertex3f(right, y, 100.0f);   // X-Y Ebene
     }
     glEnd();
     
     // Zeichne Achsen
     glColor3f(0.5f, 0.5f, 0.5f); // Hellere Farbe für Achsen
     glLineWidth(2.0f);
-
-    //TODO: Move to VBO
     glBegin(GL_LINES);
     // X-Achse
     glVertex3f(left, 0.0f, 0.0f);
     glVertex3f(right, 0.0f, 0.0f);
-    // Z-Achse
-    glVertex3f(0.0f, 0.0f, bottom);
-    glVertex3f(0.0f, 0.0f, top);
+    // Y-Achse
+    glVertex3f(0.0f, bottom, 0.0f);
+    glVertex3f(0.0f, top, 0.0f);
     glEnd();
     glLineWidth(1.0f);
 }
 
-void Orth_XZ_OpenGLWidget::updateProjectionMatrix() {
+void Orth_YZ_OpenGLWidget::updateProjectionMatrix() {
     float w = width();
     float h = height();
     
+    
     projectionMatrix.setToIdentity();
-    projectionMatrix.ortho(-w/(2.0f*zoomLevel), w/(2.0f*zoomLevel), 
-                          -1000, 1000,
+    projectionMatrix.ortho(-1000, 1000,
+                          -w/(2.0f*zoomLevel), w/(2.0f*zoomLevel), 
                           -h/(2.0f*zoomLevel), h/(2.0f*zoomLevel));
     projectionMatrix.scale(zoomLevel, zoomLevel, 1.0f);
 }
 
-QVector3D Orth_XZ_OpenGLWidget::screenToWorld(int x, int z) {
-    // Konvertiere die z-Koordinate, da OpenGL den Ursprung unten links hat
-    z = height() - z;
+QVector3D Orth_YZ_OpenGLWidget::screenToWorld(int x, int y) {
+    y = height() - y;
     
     // Normalisiere die Koordinaten auf den Bereich [-1, 1]
-    float normalizedX = (2.0f * x / width()) - 1.0f;
-    float normalizedZ = (2.0f * z / height()) - 1.0f;
+    float normalizedY = (2.0f * x / width()) - 1.0f;
+    float normalizedZ = (2.0f * y / height()) - 1.0f;
     
     // Skaliere mit dem Zoom-Faktor
-    normalizedX /= zoomLevel;
+    normalizedY /= zoomLevel;
     normalizedZ /= zoomLevel;
     
     // Skaliere mit der Fenstergröße
-    float worldX = normalizedX * (width() / 2.0f);
+    float worldY = normalizedY * (width() / 2.0f);
     float worldZ = normalizedZ * (height() / 2.0f);
     
-    return QVector3D(worldX, 0.0f, worldZ);
+    return QVector3D(0.0f, worldY, worldZ);
 }
-
