+++ src/View/mainWindow.cpp
@@ -1,11 +1,12 @@
 #include "mainWindow.h"
 #include "Controller.h"
 #include "Model.h"
+#include "View_OpenGLWidget.h"
 #include "Orth_XY_OpenGLWidget.h"
 #include "Orth_XZ_OpenGLWidget.h"
+#include "Orth_YZ_OpenGLWidget.h"
 #include <QPushButton>
 #include <QApplication>
-#include <QGridLayout>
 #include <QHBoxLayout>
 #include <QMenuBar>
 #include <QOpenGLWidget>
@@ -18,15 +19,9 @@ void mainWindow::setXYView() {
   }
   
   view = new Orth_XY_OpenGLWidget(model, controller);
-  controller->setView(static_cast<Orth_XY_OpenGLWidget*>(view));
+  controller->setView(view);
 
-//TODO: Use polymorphism instead of casting, check which view functions can be moved to parent class
-  
-  // Füge die View wieder zum Layout hinzu
-  QGridLayout *mainLayout = qobject_cast<QGridLayout*>(centralWidget()->layout());
-  if (mainLayout) {
-    mainLayout->addWidget(view, 1, 0);
-  }
+  mainLayout->addWidget(view, 1, 0);
   view->show();
  };
 
@@ -37,13 +32,22 @@ void mainWindow::setXZView() {
   }
   
   view = new Orth_XZ_OpenGLWidget(model, controller);
-  controller->setView(static_cast<Orth_XZ_OpenGLWidget*>(view));
+  controller->setView(view);
   
-  // Füge die View wieder zum Layout hinzu
-  QGridLayout *mainLayout = qobject_cast<QGridLayout*>(centralWidget()->layout());
-  if (mainLayout) {
-    mainLayout->addWidget(view, 1, 0);
+  mainLayout->addWidget(view, 1, 0);
+  view->show();
+ };
+
+void mainWindow::setYZView() {
+  if (view) {
+    view->hide();
+    delete view;
   }
+  
+  view = new Orth_YZ_OpenGLWidget(model, controller);
+  controller->setView(view);
+  
+  mainLayout->addWidget(view, 1, 0);
   view->show();
  };
 
@@ -54,7 +58,7 @@ mainWindow::mainWindow(QWidget *parent) : QMainWindow(parent) {
  model = new Model();
  controller = new Controller(model, nullptr);
  view = new Orth_XY_OpenGLWidget(model, controller);
- controller->setView(static_cast<Orth_XY_OpenGLWidget*>(view));
+ controller->setView(view);
 
 
  QWidget *centralWidget = new QWidget(this);
@@ -87,30 +91,33 @@ mainWindow::mainWindow(QWidget *parent) : QMainWindow(parent) {
 
 
  // ----- Layout -----
- QGridLayout *mainLayout = new QGridLayout(centralWidget);
+ mainLayout = new QGridLayout(centralWidget);
+ mainLayout->setContentsMargins(10, 10, 10, 10);
+ mainLayout->setSpacing(5);
 
  // -- button row to
  QHBoxLayout *buttonLayout = new QHBoxLayout();
+ buttonLayout->setAlignment(Qt::AlignLeft);
 
- mainLayout->addLayout(buttonLayout,0,0,Qt::AlignTop);
+ mainLayout->addLayout(buttonLayout, 0, 0);
 
  // -- button row bottom
  QHBoxLayout *buttonLayoutBottom = new QHBoxLayout();
+ buttonLayoutBottom->setAlignment(Qt::AlignLeft);
 
- mainLayout->addLayout(buttonLayoutBottom,1,0,Qt::AlignBottom);
+ mainLayout->addLayout(buttonLayoutBottom, 2, 0);
 
  // -- vertical stretch
- mainLayout->setColumnStretch(0, 1);
- mainLayout->setColumnStretch(1, 0);
+ mainLayout->setColumnStretch(0, 5); // View
+ mainLayout->setColumnStretch(1, 1); // Side Container
 
  // -- horizontal stretch
- mainLayout->setRowStretch(0, 0);
- mainLayout->setRowStretch(1, 1);
-
+ mainLayout->setRowStretch(0, 0);    // Top Buttons
+ mainLayout->setRowStretch(1, 10);   // View
+ mainLayout->setRowStretch(2, 0);    // Bottom Buttons
 
  // ----- Graphics Window -----
-
- mainLayout->addWidget(view);
+ mainLayout->addWidget(view, 1, 0);
 
  // ----- Layers Widget -----
  QWidget *emptyContainer = new QWidget(this);
@@ -163,4 +170,15 @@ mainWindow::mainWindow(QWidget *parent) : QMainWindow(parent) {
  QObject::connect(XZView, SIGNAL(clicked()), this, SLOT(setXZView()));
 
  buttonLayoutBottom->addWidget(XZView);
+
+ QPushButton *YZView = new QPushButton("YZ View",  this);
+ YZView->setText("YZ View");
+ YZView->setToolTip("YZ View");
+ YZView->show();
+
+ QObject::connect(YZView, SIGNAL(clicked()), this, SLOT(setYZView()));
+
+ buttonLayoutBottom->addWidget(YZView);
+
+ buttonLayoutBottom->addStretch(1);
 }
